RD-2008-06-26-0: Enhancem.: - removed SIMULATE_EXECDS param                             - added SIMULATE_JOBS param                   Review:   JG

+++ source/daemons/execd/reaper_execd.c
@@ -988,24 +988,26 @@ lListElem *jr
       }
 
       if (!pe_task_id_str) {
-         job_remove_spool_file(job_id, ja_task_id, NULL, SPOOL_WITHIN_EXECD);
+         if (!mconf_get_simulate_jobs()) {
+            job_remove_spool_file(job_id, ja_task_id, NULL, SPOOL_WITHIN_EXECD);
 
-         if (!JOB_TYPE_IS_BINARY(lGetUlong(jep, JB_type)) &&
-             lGetString(jep, JB_exec_file)) {
-            int task_number = 0;
-            lListElem *tmp_job = NULL;
+            if (!JOB_TYPE_IS_BINARY(lGetUlong(jep, JB_type)) &&
+                lGetString(jep, JB_exec_file)) {
+               int task_number = 0;
+               lListElem *tmp_job = NULL;
 
-            /* it is possible to remove the exec_file if
-               less than one task of a job is running */
-            tmp_job = lGetElemUlongFirst(Master_Job_List, JB_job_number, job_id, &iterator);
-            while(tmp_job != NULL) {
-               task_number++;
-               tmp_job = lGetElemUlongNext(Master_Job_List, JB_job_number, job_id, &iterator);
-            }
-            
-            if (task_number <= 1) {
-               DPRINTF(("unlinking script file %s\n", lGetString(jep, JB_exec_file)));
-               unlink(lGetString(jep, JB_exec_file));
+               /* it is possible to remove the exec_file if
+                  less than one task of a job is running */
+               tmp_job = lGetElemUlongFirst(Master_Job_List, JB_job_number, job_id, &iterator);
+               while(tmp_job != NULL) {
+                  task_number++;
+                  tmp_job = lGetElemUlongNext(Master_Job_List, JB_job_number, job_id, &iterator);
+               }
+               
+               if (task_number <= 1) {
+                  DPRINTF(("unlinking script file %s\n", lGetString(jep, JB_exec_file)));
+                  unlink(lGetString(jep, JB_exec_file));
+               }
             }
          }
       } else {
@@ -1083,7 +1085,9 @@ lListElem *jr
          }
 
          /* job */
-         job_remove_spool_file(job_id, ja_task_id, NULL, SPOOL_WITHIN_EXECD); 
+         if (!mconf_get_simulate_jobs()) {
+            job_remove_spool_file(job_id, ja_task_id, NULL, SPOOL_WITHIN_EXECD); 
+         }
 
          /* active dir */
          if (!mconf_get_keep_active() && !getenv("SGE_KEEP_ACTIVE")) {
@@ -1264,7 +1268,9 @@ int startup
       exited and there is no need for ps-commands.
 
    */
-   if (!Master_Job_List || !lGetNumberOfElem(Master_Job_List) || !lost_children) {
+   if (mconf_get_simulate_jobs() ||
+       !Master_Job_List || !lGetNumberOfElem(Master_Job_List) ||
+       !lost_children) {
       if (lost_children) {
          INFO((SGE_EVENT, MSG_SHEPHERD_NOOLDJOBSATSTARTUP));
          lost_children = 0;
